refactoring and fix some problems
