+++ src/hooks/useInstallmentManagement.tsx
@@ -1,102 +1,147 @@
-import { useEffect, useState, useRef } from 'react';
+/**
+ * Hook for installment payment management
+ * Improved version using reducer pattern
+ */
+import { useReducer, useCallback, useEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  useFieldArray,
-  UseFormGetValues,
-  UseFormTrigger,
-  Path,
-  FieldValues,
-  FieldArrayPath,
-  PathValue,
-  Control,
-  UseFormSetValue
-} from 'react-hook-form';
+import { useFieldArray, Path, FieldValues, PathValue } from 'react-hook-form';
 import { createAmountValidator } from '../utils/fieldValidation';
 import { formatDate, moneyFormat } from '../utils/formatters';
+import {
+  Installment,
+  InstallmentManagementProps,
+  InstallmentManagementResult,
+  InstallmentValidators
+} from '../models/paymentTypes';
 
-// Definizione del tipo per i validatori delle rate
-export type InstallmentValidators = {
-  amount: Record<string, unknown>;
-  dueDate: {
-    required: string | boolean;
-  };
-};
-
-export type InstallmentData = {
-  amount: string;
-  dueDate: string | null;
-  isMultibeneficiary?: boolean;
-  beneficiaries?: Array<{
-    entityName: string;
-    amount: string;
-    taxCode: string;
-    iban: string;
-    postalAccount: string;
-    taxonomyCode: string;
-  }>;
-  id?: string;
-  isNew?: boolean;
-};
-
-export type InstallmentFormValues = {
-  installments: Array<InstallmentData>;
-};
-
-type InstallmentFieldPath<T extends FieldValues> = FieldArrayPath<T>;
-
-type UseInstallmentManagementProps<T extends FieldValues> = {
-  readonly control: Control<T>;
-  readonly fieldNamePrefix: InstallmentFieldPath<T>;
-  readonly isSubmitted: boolean;
-  readonly getValues: UseFormGetValues<T>;
-  readonly setValue: UseFormSetValue<T>;
-  readonly trigger: UseFormTrigger<T>;
-  readonly flagMandatoryDueDate?: boolean;
-  readonly onInstallmentsChange?: (
-    installments: Array<InstallmentData>,
-    totalAmount: string
-  ) => void;
+/**
+ * Internal hook state
+ */
+type InstallmentState = {
+  /** Registry of installments existing before submit */
+  existingInstallments: Record<string, boolean>;
+  /** Flag indicating if form was submitted */
+  wasSubmitted: boolean;
+  /** Flag indicating if initialization is in progress */
+  isInitializing: boolean;
+  /** Flag indicating if initialization is complete */
+  hasInitialized: boolean;
+  /** Last calculated total amount to prevent unnecessary updates */
+  lastTotalAmount: string;
+  /** Flag to prevent recursive updates */
+  isUpdating: boolean;
 };
 
 /**
- * Hook personalizzato per la gestione delle rate
- * Gestisce l'aggiunta, rimozione e validazione delle rate
+ * Actions for the reducer
  */
-export function useInstallmentManagement<T extends FieldValues>({
-  control,
-  fieldNamePrefix,
-  isSubmitted,
-  getValues,
-  trigger,
-  flagMandatoryDueDate = true,
-  onInstallmentsChange
-}: UseInstallmentManagementProps<T>) {
-  // ===== CONSTANTS =====
+type InstallmentAction =
+  | { type: 'MARK_SUBMITTED' }
+  | { type: 'SET_EXISTING_INSTALLMENTS'; installments: Record<string, boolean> }
+  | { type: 'START_INITIALIZING' }
+  | { type: 'FINISH_INITIALIZING' }
+  | { type: 'MARK_INITIALIZED' }
+  | { type: 'SET_LAST_TOTAL_AMOUNT'; amount: string }
+  | { type: 'START_UPDATING' }
+  | { type: 'FINISH_UPDATING' };
+
+/**
+ * Reducer to centrally manage installment state
+ */
+function installmentReducer(
+  state: InstallmentState,
+  action: InstallmentAction
+): InstallmentState {
+  switch (action.type) {
+    case 'MARK_SUBMITTED':
+      return {
+        ...state,
+        wasSubmitted: true
+      };
+    case 'SET_EXISTING_INSTALLMENTS':
+      return {
+        ...state,
+        existingInstallments: action.installments
+      };
+    case 'START_INITIALIZING':
+      return {
+        ...state,
+        isInitializing: true
+      };
+    case 'FINISH_INITIALIZING':
+      return {
+        ...state,
+        isInitializing: false
+      };
+    case 'MARK_INITIALIZED':
+      return {
+        ...state,
+        hasInitialized: true
+      };
+    case 'SET_LAST_TOTAL_AMOUNT':
+      return {
+        ...state,
+        lastTotalAmount: action.amount
+      };
+    case 'START_UPDATING':
+      return {
+        ...state,
+        isUpdating: true
+      };
+    case 'FINISH_UPDATING':
+      return {
+        ...state,
+        isUpdating: false
+      };
+    default:
+      return state;
+  }
+}
+
+/**
+ * Improved hook for installment management
+ * Uses useReducer for more predictable state management
+ */
+export function useInstallmentManagement<T extends FieldValues>(
+  props: InstallmentManagementProps<T>
+): InstallmentManagementResult {
+  const {
+    control,
+    fieldNamePrefix,
+    isSubmitted,
+    getValues,
+    trigger,
+    flagMandatoryDueDate = true,
+    onInstallmentsChange
+  } = props;
+
   const MIN_INSTALLMENTS = 2;
   const MAX_INSTALLMENTS = 12;
   const { t } = useTranslation();
+  const [state, dispatch] = useReducer(installmentReducer, {
+    existingInstallments: {},
+    wasSubmitted: false,
+    isInitializing: false,
+    hasInitialized: false,
+    lastTotalAmount: '',
+    isUpdating: false
+  });
 
-  // ===== STATE & REFS =====
-  const [existingInstallments, setExistingInstallments] = useState<
-    Record<string, boolean>
-  >({});
-  const wasSubmittedRef = useRef(false);
-  const isInitializingRef = useRef(false);
-  const hasInitializedRef = useRef(false);
-  // Aggiungiamo un ref per memorizzare l'ultimo totale
-  const lastTotalAmountRef = useRef('');
-  // Ref per evitare aggiornamenti ripetuti
-  const isUpdatingRef = useRef(false);
+  // Refs for compatibility with original API
+  const wasSubmittedRef = useRef(state.wasSubmitted);
+  const isInitializingRef = useRef(state.isInitializing);
 
-  // ===== FIELD ARRAY =====
+  // Update refs when state changes
+  useEffect(() => {
+    wasSubmittedRef.current = state.wasSubmitted;
+    isInitializingRef.current = state.isInitializing;
+  }, [state.wasSubmitted, state.isInitializing]);
+
   const { fields, append, remove } = useFieldArray({
     control,
     name: fieldNamePrefix
   });
 
-  // ===== VALIDATORS =====
-  // Creazione diretta dei validatori senza useMemo
-  // L'obbligatorietà della data di scadenza dipende dal flag flagMandatoryDueDate
   const validators: InstallmentValidators = {
     amount: createAmountValidator(t),
     dueDate: {
@@ -106,17 +151,19 @@ export function useInstallmentManagement<T extends FieldValues>({
     }
   };
 
-  // ===== UTILITY FUNCTIONS =====
-  // Ottiene i dati attuali di tutte le rate
-  function getInstallmentsData(): Array<InstallmentData> {
+  /**
+   * Get current data for all installments
+   */
+  const getInstallmentsData = useCallback((): Array<Installment> => {
     return fields.map((field, index) => {
       const installmentData = getValues(
         `${fieldNamePrefix}.${index}` as Path<T>
       );
-      // Se la data è un oggetto Date, formattala prima di restituirla
+
+      // If date is a Date object, format it before returning
       const dueDate = installmentData?.dueDate;
 
-      // Ottieni l'importo e formattalo se presente
+      // Get amount and format it if present
       const amount = installmentData?.amount;
       const formattedAmount = amount
         ? moneyFormat(parseFloat(amount) * 100)
@@ -126,21 +173,24 @@ export function useInstallmentManagement<T extends FieldValues>({
 
       return {
         ...installmentData,
-        // Formatta la data solo se è un oggetto Date e se è valida
+        // Format date only if it's a Date object and valid
         dueDate:
           dueDate instanceof Date && !isNaN(dueDate.getTime())
             ? formatDate(dueDate.toISOString())
             : dueDate,
-        // Usa il valore formattato per l'importo
+        // Use formatted value for amount
         amount: formattedAmount,
         id: field.id,
-        isNew: !!wasSubmittedRef.current && !existingInstallments[field.id]
-      } as InstallmentData;
+        isNew:
+          !!wasSubmittedRef.current && !state.existingInstallments[field.id]
+      } as Installment;
     });
-  }
+  }, [fields, getValues, fieldNamePrefix, state.existingInstallments]);
 
-  // Calcola l'importo totale sommando tutte le rate
-  function calculateTotalAmount(): string {
+  /**
+   * Calculate total amount by summing all installments
+   */
+  const calculateTotalAmount = useCallback((): string => {
     return fields
       .reduce((total, _, index) => {
         const installmentData = getValues(
@@ -151,24 +201,24 @@ export function useInstallmentManagement<T extends FieldValues>({
         return total + amountValue;
       }, 0)
       .toFixed(2);
-  }
+  }, [fields, getValues, fieldNamePrefix]);
 
-  // ===== INSTALLMENT MANAGEMENT =====
-  // Aggiunge una nuova rata se il limite massimo non è stato raggiunto
-  function addInstallment() {
+  /**
+   * Add a new installment
+   */
+  const addInstallment = useCallback(() => {
     if (fields.length < MAX_INSTALLMENTS) {
-      // Crea una nuova rata semplice con data di scadenza vuota
-      const newInstallment: InstallmentData = {
+      // Create a new simple installment with empty due date
+      const newInstallment: Installment = {
         amount: '',
-        dueDate: null, // Data di scadenza sempre vuota per default
+        dueDate: null, // Due date always empty by default
         isMultibeneficiary: false
       };
 
-      append(
-        newInstallment as unknown as PathValue<T, InstallmentFieldPath<T>>
-      );
+      // Add the new installment to the form
+      append(newInstallment as unknown as PathValue<T, typeof fieldNamePrefix>);
 
-      // Aggiorna gli importi e notifica i cambiamenti
+      // Update amounts and notify changes
       setTimeout(() => {
         const newTotalAmount = calculateTotalAmount();
         if (onInstallmentsChange) {
@@ -177,32 +227,52 @@ export function useInstallmentManagement<T extends FieldValues>({
         }
       }, 0);
     }
-  }
+  }, [
+    fields.length,
+    MAX_INSTALLMENTS,
+    append,
+    fieldNamePrefix,
+    calculateTotalAmount,
+    onInstallmentsChange,
+    getInstallmentsData
+  ]);
 
-  // Rimuove una rata all'indice specificato
-  function removeInstallment(index: number) {
-    // Non permettere la rimozione se rimangono solo le due rate minime richieste
-    if (fields.length <= MIN_INSTALLMENTS) {
-      return;
-    }
-
-    remove(index);
-
-    // Aggiorna gli importi e notifica i cambiamenti
-    setTimeout(() => {
-      const newTotalAmount = calculateTotalAmount();
-      if (onInstallmentsChange) {
-        const currentInstallments = getInstallmentsData();
-        onInstallmentsChange(currentInstallments, newTotalAmount);
+  /**
+   * Remove installment at specified index
+   */
+  const removeInstallment = useCallback(
+    (index: number) => {
+      // Don't allow removal if only minimum required installments remain
+      if (fields.length <= MIN_INSTALLMENTS) {
+        return;
       }
-    }, 0);
-  }
 
-  // ===== EFFECT HOOKS =====
-  // Registra le rate esistenti al primo submit
+      // Remove installment from form
+      remove(index);
+
+      // Update amounts and notify changes
+      setTimeout(() => {
+        const newTotalAmount = calculateTotalAmount();
+        if (onInstallmentsChange) {
+          const currentInstallments = getInstallmentsData();
+          onInstallmentsChange(currentInstallments, newTotalAmount);
+        }
+      }, 0);
+    },
+    [
+      fields.length,
+      MIN_INSTALLMENTS,
+      remove,
+      calculateTotalAmount,
+      onInstallmentsChange,
+      getInstallmentsData
+    ]
+  );
+
+  // Register existing installments on first submit
   useEffect(() => {
-    if (isSubmitted && !wasSubmittedRef.current) {
-      // Prima volta che viene fatto submit - memorizziamo lo stato attuale delle rate
+    if (isSubmitted && !state.wasSubmitted) {
+      // Store current state of installments
       const currentInstallments = fields.reduce<Record<string, boolean>>(
         (acc, field) => {
           acc[field.id] = true;
@@ -210,87 +280,114 @@ export function useInstallmentManagement<T extends FieldValues>({
         },
         {}
       );
-      setExistingInstallments(currentInstallments);
-      wasSubmittedRef.current = true;
-    }
-  }, [isSubmitted, fields]);
 
-  // Aggiorna validazione quando cambiano gli importi
+      // Update state
+      dispatch({
+        type: 'SET_EXISTING_INSTALLMENTS',
+        installments: currentInstallments
+      });
+      dispatch({ type: 'MARK_SUBMITTED' });
+    }
+  }, [isSubmitted, fields, state.wasSubmitted]);
+
+  // Update validation when amounts change
   useEffect(() => {
-    if (wasSubmittedRef.current) {
+    if (state.wasSubmitted) {
       fields.forEach((field, index) => {
-        if (existingInstallments[field.id]) {
+        if (state.existingInstallments[field.id]) {
           trigger(`${fieldNamePrefix}.${index}` as Path<T>);
         }
       });
     }
-  }, [trigger, fieldNamePrefix, fields, existingInstallments]);
+  }, [
+    trigger,
+    fieldNamePrefix,
+    fields,
+    state.wasSubmitted,
+    state.existingInstallments
+  ]);
 
-  // Inizializza le prime due rate se non ce ne sono
+  // Initialize first two installments if none exist
   useEffect(() => {
-    if (fields.length === 0 && !hasInitializedRef.current) {
-      hasInitializedRef.current = true;
-      isInitializingRef.current = true;
-
-      // Crea due rate iniziali semplici con date di scadenza vuote
-      const firstInstallment: InstallmentData = {
-        amount: '',
-        dueDate: null, // Data di scadenza vuota
-        isMultibeneficiary: false
-      };
-
-      const secondInstallment: InstallmentData = {
-        amount: '',
-        dueDate: null, // Data di scadenza vuota
-        isMultibeneficiary: false
-      };
-
-      // Aggiungi le due rate
-      append(
-        firstInstallment as unknown as PathValue<T, InstallmentFieldPath<T>>
-      );
-      append(
-        secondInstallment as unknown as PathValue<T, InstallmentFieldPath<T>>
-      );
-
-      isInitializingRef.current = false;
+    // Add explicit check on hasInitialized value
+    // to ensure multiple installments aren't added
+    if (fields.length === 0 && !state.hasInitialized) {
+      // First ensure no other initializations can occur
+      dispatch({ type: 'MARK_INITIALIZED' });
+      // Then start initialization phase
+      dispatch({ type: 'START_INITIALIZING' });
+      try {
+        // Check if there are already data in the form
+        const formValue = getValues(fieldNamePrefix as unknown as Path<T>);
+        if (Array.isArray(formValue) && formValue.length > 0) {
+          dispatch({ type: 'FINISH_INITIALIZING' });
+          return;
+        }
+        // Create two initial simple installments with empty due dates
+        const firstInstallment: Installment = {
+          amount: '',
+          dueDate: null, // Empty due date
+          isMultibeneficiary: false
+        };
+        const secondInstallment: Installment = {
+          amount: '',
+          dueDate: null, // Empty due date
+          isMultibeneficiary: false
+        };
+        // Use append sequentially and synchronously
+        append(
+          firstInstallment as unknown as PathValue<T, typeof fieldNamePrefix>
+        );
+        append(
+          secondInstallment as unknown as PathValue<T, typeof fieldNamePrefix>
+        );
+      } catch (error) {
+        console.error('Error during initialization:', error);
+      } finally {
+        dispatch({ type: 'FINISH_INITIALIZING' });
+      }
     }
-  }, [fields.length, append]);
+  }, [fields.length, append, fieldNamePrefix, state.hasInitialized, getValues]);
 
-  // Notifica le modifiche alle rate e calcola il totale
+  // Notify changes and calculate total - only for existing installments
   useEffect(() => {
-    // Verifichiamo se siamo già in fase di aggiornamento per evitare cicli
-    if (isUpdatingRef.current) {
+    // Check if we're already updating to avoid cycles
+    // or if we're initializing (in which case we don't want to notify)
+    if (state.isUpdating || state.isInitializing) {
       return;
     }
 
-    if (
-      onInstallmentsChange &&
-      fields.length > 0 &&
-      !isInitializingRef.current
-    ) {
-      // Calcola il nuovo totale
+    // Notify only if there are installments and we're not initializing
+    if (onInstallmentsChange && fields.length > 0) {
+      // Calculate new total
       const totalAmount = calculateTotalAmount();
-
-      // Confronta con l'ultimo totale memorizzato per evitare aggiornamenti inutili
-      if (totalAmount !== lastTotalAmountRef.current) {
-        // Imposta il flag di aggiornamento per evitare chiamate ricorsive
-        isUpdatingRef.current = true;
-        // Aggiorna il valore di riferimento
-        lastTotalAmountRef.current = totalAmount;
-        // Esegui la callback
+      // Compare with last stored total to avoid unnecessary updates
+      if (totalAmount !== state.lastTotalAmount) {
+        // Set update flag to prevent recursive calls
+        dispatch({ type: 'START_UPDATING' });
+        // Update reference value
+        dispatch({ type: 'SET_LAST_TOTAL_AMOUNT', amount: totalAmount });
+        // Execute callback
         const currentInstallments = getInstallmentsData();
         onInstallmentsChange(currentInstallments, totalAmount);
-        // Reimposta il flag dopo il completamento
-        isUpdatingRef.current = false;
+        // Reset flag after completion
+        dispatch({ type: 'FINISH_UPDATING' });
       }
     }
-  }, [fields.length, onInstallmentsChange]);
+  }, [
+    fields.length,
+    onInstallmentsChange,
+    calculateTotalAmount,
+    getInstallmentsData,
+    state.isInitializing,
+    state.isUpdating,
+    state.lastTotalAmount
+  ]);
 
   return {
     fields,
     validators,
-    existingInstallments,
+    existingInstallments: state.existingInstallments,
     MIN_INSTALLMENTS,
     MAX_INSTALLMENTS,
     wasSubmittedRef,
