+++ src/routes/DebtPositionCreateWizard/components/Beneficiary/BeneficiaryField.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from 'react-i18next';
-import React, { forwardRef, useImperativeHandle } from 'react';
+import React, { useImperativeHandle } from 'react';
 import {
   Control,
   UseFormGetValues,
@@ -22,18 +22,20 @@ import {
 import { useBeneficiaryManagement } from '../../../../hooks/useBeneficiaryManagement';
 import { useInstallmentBeneficiaryManagement } from '../../../../hooks/useInstallmentBeneficiaryManagement';
 
-type BeneficiaryFieldPath = FieldArrayPath<FieldValues>;
+// ===== TYPES =====
+type BeneficiaryField = Record<string, unknown> & { id: string };
 
+// Type for functions exposed by ref
 export type BeneficiaryFieldRef = {
   resetAllBeneficiaries?: () => void;
 };
 
-type BeneficiaryFieldProps<T extends FieldValues> = {
+type BeneficiaryFieldProps<T extends FieldValues = FieldValues> = {
   readonly control: Control<T>;
   readonly isSubmitted: boolean;
   readonly errors: FieldErrors<T>;
   readonly totalAmount: string;
-  readonly fieldNamePrefix: BeneficiaryFieldPath;
+  readonly fieldNamePrefix: string;
   readonly disabled?: boolean;
   readonly setValue: UseFormSetValue<T>;
   readonly getValues: UseFormGetValues<T>;
@@ -53,186 +55,238 @@ type BeneficiaryFieldProps<T extends FieldValues> = {
 };
 
 /**
- * Main component for beneficiary management
+ * Main component for beneficiary management - Improved version with useReducer
  * Allows adding, removing and validating up to 4 beneficiaries
+ * Maintains the same UI but uses reducer hooks for better state management
  */
-const BeneficiaryField = forwardRef(
-  (
-    props: BeneficiaryFieldProps<FieldValues>,
-    ref: React.ForwardedRef<BeneficiaryFieldRef>
-  ) => {
-    const {
-      control,
-      isSubmitted,
-      errors,
-      totalAmount,
-      fieldNamePrefix,
-      disabled = false,
-      setValue,
-      getValues,
-      trigger,
-      onToggleMultibeneficiary,
-      onBeneficiariesChange,
-      isInsideInstallment = false,
-      installmentIndex,
-      installmentsFieldNamePrefix
-    } = props;
+interface BeneficiaryFieldComponent {
+  <T extends FieldValues>(
+    props: BeneficiaryFieldProps<T> & {
+      ref?: React.ForwardedRef<BeneficiaryFieldRef>;
+    }
+  ): React.ReactElement;
+  displayName?: string;
+}
 
-    const { t } = useTranslation();
+const InternalBeneficiaryField = <T extends FieldValues>(
+  props: BeneficiaryFieldProps<T>,
+  ref: React.ForwardedRef<BeneficiaryFieldRef>
+) => {
+  const {
+    control,
+    isSubmitted,
+    errors,
+    totalAmount,
+    fieldNamePrefix,
+    disabled = false,
+    setValue,
+    getValues,
+    trigger,
+    onToggleMultibeneficiary,
+    onBeneficiariesChange,
+    isInsideInstallment = false,
+    installmentIndex,
+    installmentsFieldNamePrefix
+  } = props;
 
-    const beneficiaryManager =
-      isInsideInstallment &&
-      installmentIndex !== undefined &&
-      installmentsFieldNamePrefix
-        ? useInstallmentBeneficiaryManagement<FieldValues>({
-            control,
-            index: installmentIndex,
-            installmentsFieldNamePrefix,
-            isSubmitted,
-            getValues,
-            setValue,
-            trigger,
-            onToggleMultibeneficiary
-          })
-        : useBeneficiaryManagement<FieldValues>({
-            control,
-            fieldNamePrefix,
-            isSubmitted,
-            getValues,
-            trigger,
-            totalAmount,
-            onToggleMultibeneficiary,
-            onBeneficiariesChange
-          });
+  const { t } = useTranslation();
 
-    const {
-      fields,
-      validators,
-      fieldValidators,
-      MAX_BENEFICIARIES,
-      existingBeneficiaries,
-      wasSubmittedRef,
-      addBeneficiary,
-      removeBeneficiary,
-      resetAllBeneficiaries
-    } = beneficiaryManager;
+  const beneficiaryManager =
+    isInsideInstallment &&
+    installmentIndex !== undefined &&
+    installmentsFieldNamePrefix
+      ? useInstallmentBeneficiaryManagement<T>({
+          control,
+          index: installmentIndex,
+          installmentsFieldNamePrefix,
+          isSubmitted,
+          getValues,
+          setValue,
+          trigger,
+          onToggleMultibeneficiary
+        })
+      : useBeneficiaryManagement<T>({
+          control,
+          fieldNamePrefix: fieldNamePrefix as FieldArrayPath<T>,
+          isSubmitted,
+          getValues,
+          trigger,
+          totalAmount,
+          onToggleMultibeneficiary,
+          onBeneficiariesChange
+        });
 
-    useImperativeHandle(ref, () => ({
-      resetAllBeneficiaries
-    }));
+  const {
+    fields,
+    validators,
+    fieldValidators,
+    MAX_BENEFICIARIES,
+    existingBeneficiaries,
+    wasSubmittedRef,
+    addBeneficiary,
+    removeBeneficiary,
+    resetAllBeneficiaries
+  } = beneficiaryManager;
 
-    const renderAddBeneficiaryButton = (index: number): JSX.Element | null => {
-      if (index === fields.length - 1 && fields.length < MAX_BENEFICIARIES) {
-        return (
-          <>
-            <Divider sx={{ mt: 2, mb: 1 }} />
-            <Box>
-              <Button
-                startIcon={<Add />}
-                onClick={addBeneficiary}
-                sx={{ mt: 1 }}
-                variant="text"
-                color="primary"
-                disabled={fields.length >= MAX_BENEFICIARIES}
-              >
-                {t('debtPositionCreateWizard.step3.beneficiary.addBeneficiary')}
-              </Button>
-            </Box>
-          </>
-        );
-      }
-      return null;
-    };
+  useImperativeHandle(ref, () => ({
+    resetAllBeneficiaries
+  }));
 
-    const createValidationContext = (
-      field: Record<'id', string>,
-      index: number
-    ): BeneficiaryValidationContext<FieldValues> => {
-      return {
-        id: field.id,
-        index,
-        isSubmitted,
-        wasSubmittedRef,
-        existingBeneficiaries,
-        errors,
-        fieldNamePrefix,
-        getValues,
-        t
-      };
-    };
-
-    const renderBeneficiary = (field: Record<'id', string>, index: number) => {
-      const validationContext = createValidationContext(field, index);
-
+  /**
+   * Renders the button to add a new beneficiary
+   * Only visible on the last beneficiary and if the maximum limit hasn't been reached
+   */
+  const renderAddBeneficiaryButton = (index: number): JSX.Element | null => {
+    if (index === fields.length - 1 && fields.length < MAX_BENEFICIARIES) {
       return (
-        <Paper
-          key={field.id}
-          sx={{
-            p: 2,
-            mb: 2,
-            position: 'relative'
-          }}
-        >
-          <BeneficiaryHeader index={index} t={t} onRemove={removeBeneficiary} />
-
-          <Grid container spacing={2}>
-            <BeneficiaryIdentityFields
-              control={control}
-              index={index}
-              fieldNamePrefix={fieldNamePrefix}
-              validationContext={validationContext}
-              disabled={disabled}
-              t={t}
-            />
-
-            <BeneficiaryAmountFields
-              control={control}
-              index={index}
-              fieldNamePrefix={fieldNamePrefix}
-              validationContext={validationContext}
-              disabled={disabled}
-              fields={fields}
-              validators={validators}
-              trigger={trigger}
-              t={t}
-            />
-
-            <BeneficiaryPaymentFields
-              control={control}
-              index={index}
-              fieldNamePrefix={fieldNamePrefix}
-              validationContext={validationContext}
-              disabled={disabled}
-              getValues={getValues}
-              trigger={trigger}
-              errors={errors}
-              fieldValidators={fieldValidators}
-              t={t}
-            />
-
-            <BeneficiaryClassificationFields
-              control={control}
-              index={index}
-              fieldNamePrefix={fieldNamePrefix}
-              validationContext={validationContext}
-              disabled={disabled}
-              t={t}
-            />
-          </Grid>
-
-          {renderAddBeneficiaryButton(index)}
-        </Paper>
+        <>
+          <Divider sx={{ mt: 2, mb: 1 }} />
+          <Box>
+            <Button
+              startIcon={<Add />}
+              onClick={addBeneficiary}
+              sx={{ mt: 1 }}
+              variant="text"
+              color="primary"
+              disabled={fields.length >= MAX_BENEFICIARIES}
+            >
+              {t('debtPositionCreateWizard.step3.beneficiary.addBeneficiary')}
+            </Button>
+          </Box>
+        </>
       );
+    }
+    return null;
+  };
+
+  /**
+   * Creates validation context for a beneficiary
+   */
+  const createValidationContext = (
+    field: BeneficiaryField,
+    index: number
+  ): BeneficiaryValidationContext<T> => {
+    return {
+      id: field.id,
+      index,
+      isSubmitted,
+      wasSubmittedRef,
+      existingBeneficiaries,
+      errors,
+      fieldNamePrefix,
+      getValues,
+      t
     };
+  };
+
+  /**
+   * Renders a single beneficiary with its field groups
+   */
+  const renderBeneficiary = (field: BeneficiaryField, index: number) => {
+    const validationContext = createValidationContext(field, index);
 
     return (
-      <Box>
-        {fields.map((field: Record<'id', string>, index: number) =>
-          renderBeneficiary(field, index)
-        )}
-      </Box>
+      <Paper
+        key={field.id}
+        sx={{
+          p: 2,
+          mb: 2,
+          position: 'relative'
+        }}
+      >
+        <BeneficiaryHeader index={index} t={t} onRemove={removeBeneficiary} />
+
+        <Grid container spacing={2}>
+          <BeneficiaryIdentityFields
+            control={control}
+            index={index}
+            fieldNamePrefix={fieldNamePrefix}
+            validationContext={validationContext}
+            disabled={disabled}
+            t={t}
+          />
+
+          <BeneficiaryAmountFields
+            control={control}
+            index={index}
+            fieldNamePrefix={fieldNamePrefix}
+            validationContext={validationContext}
+            disabled={disabled}
+            fields={fields as Array<Record<'id', string>>}
+            validators={
+              validators as unknown as {
+                isValidTotalAmount: () => boolean;
+                isSingleBeneficiaryAmountValid: (
+                  hasSingleBeneficiary: boolean
+                ) => boolean;
+                validateTotalAmount: () => string | true;
+                validateSingleBeneficiary: (
+                  amount: string,
+                  fieldsLength: number
+                ) => string | true;
+                isBeneficiaryAmountValid: (
+                  index: number,
+                  hasSingleBeneficiary: boolean
+                ) => boolean;
+              }
+            }
+            trigger={trigger}
+            t={t}
+          />
+
+          <BeneficiaryPaymentFields
+            control={control}
+            index={index}
+            fieldNamePrefix={fieldNamePrefix}
+            validationContext={validationContext}
+            disabled={disabled}
+            getValues={getValues}
+            trigger={trigger}
+            errors={errors}
+            fieldValidators={
+              fieldValidators as unknown as {
+                validateBeneficiaryTaxCode: (
+                  value: string
+                ) => string | undefined;
+                validateIBAN: (value: string) => string | undefined;
+                validatePostalAccount: (value: string) => string | undefined;
+                validatePaymentMethod: (
+                  iban: string,
+                  postalAccount: string
+                ) => string | undefined;
+              }
+            }
+            t={t}
+          />
+
+          <BeneficiaryClassificationFields
+            control={control}
+            index={index}
+            fieldNamePrefix={fieldNamePrefix}
+            validationContext={validationContext}
+            disabled={disabled}
+            t={t}
+          />
+        </Grid>
+
+        {renderAddBeneficiaryButton(index)}
+      </Paper>
     );
-  }
-);
+  };
+
+  return (
+    <Box>
+      {fields.map((field, index) =>
+        renderBeneficiary(field as BeneficiaryField, index)
+      )}
+    </Box>
+  );
+};
+
+const BeneficiaryField = React.forwardRef(
+  InternalBeneficiaryField
+) as unknown as BeneficiaryFieldComponent;
+
+BeneficiaryField.displayName = 'BeneficiaryField';
 
 export default BeneficiaryField;
