--- src/routes/DebtPositionCreateWizard/components/Installment/InstallmentItem.test.tsx
@@ -1,554 +1,0 @@
-import { render, screen, fireEvent } from '@testing-library/react';
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-import {
-  Control,
-  FieldErrors,
-  UseFormGetValues,
-  UseFormSetValue,
-  UseFormTrigger,
-  Path
-} from 'react-hook-form';
-import InstallmentItem from './InstallmentItem';
-import { InstallmentValidators } from '../../../../hooks/useInstallmentManagement';
-import { ReactNode } from 'react';
-
-// Mock delle dipendenze e utility
-vi.mock('../../../utils/formatters', () => ({
-  moneyFormat: vi.fn(
-    (value, maxDecimals) => `€ ${(value / 100).toFixed(maxDecimals || 2)}`
-  )
-}));
-
-vi.mock('react-i18next', () => ({
-  useTranslation: () => ({
-    t: (key: string) => key
-  })
-}));
-
-// Mock del console.log che viene usato nel componente
-console.log = vi.fn();
-
-// Mock di react-hook-form
-vi.mock('react-hook-form', async () => {
-  const actual = await vi.importActual('react-hook-form');
-  return {
-    ...actual,
-    Controller: ({
-      render
-    }: {
-      name?: string;
-      control?: Record<string, unknown>;
-      rules?: Record<string, unknown>;
-      render: (props: { field: Record<string, unknown> }) => JSX.Element;
-    }) => {
-      const field = {
-        value: '',
-        onChange: vi.fn(),
-        onBlur: vi.fn(),
-        ref: { current: null }
-      };
-      return render({ field });
-    }
-  };
-});
-
-vi.mock('@mui/x-date-pickers/DatePicker', () => ({
-  DatePicker: vi.fn(({ label, value, onChange, slotProps }) => (
-    <div data-testid="date-picker">
-      <label>{label}</label>
-      <input
-        type="date"
-        value={value ? value.toISOString().split('T')[0] : ''}
-        onChange={(e) => onChange(new Date(e.target.value))}
-        required={slotProps?.textField?.required}
-        data-error={slotProps?.textField?.error}
-        data-helper-text={slotProps?.textField?.helperText}
-      />
-      {slotProps?.field?.clearable && (
-        <button
-          data-testid="clear-date-button"
-          onClick={() => slotProps.field.onClear()}
-        >
-          Clear
-        </button>
-      )}
-    </div>
-  ))
-}));
-
-// Mock per Material UI
-vi.mock('@mui/material', () => {
-  return {
-    Box: ({
-      children
-    }: {
-      children: ReactNode;
-      sx?: Record<string, unknown>;
-    }) => <div data-testid="mui-box">{children}</div>,
-    Typography: ({ children }: { children: ReactNode; variant?: string }) => (
-      <span data-testid="mui-typography">{children}</span>
-    ),
-    TextField: ({
-      label,
-      required,
-      disabled,
-      error,
-      helperText,
-      value,
-      onChange,
-      onBlur
-    }: {
-      label?: string;
-      required?: boolean;
-      disabled?: boolean;
-      error?: boolean;
-      helperText?: string;
-      value?: string;
-      onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-      onBlur?: (e: React.FocusEvent<HTMLInputElement>) => void;
-      fullWidth?: boolean;
-      InputProps?: Record<string, unknown>;
-      size?: string;
-    }) => (
-      <div data-testid="mui-textfield">
-        <label>{label}</label>
-        <input
-          value={value}
-          onChange={onChange}
-          onBlur={onBlur}
-          required={required}
-          disabled={disabled}
-          data-error={error}
-          data-testid="textfield-input"
-        />
-        {helperText && <div data-testid="helper-text">{helperText}</div>}
-      </div>
-    ),
-    Grid: ({
-      children
-    }: {
-      children: ReactNode;
-      container?: boolean;
-      item?: boolean;
-      xs?: number;
-    }) => <div data-testid="mui-grid">{children}</div>,
-    IconButton: ({
-      children,
-      onClick
-    }: {
-      children: ReactNode;
-      onClick?: () => void;
-      size?: string;
-      sx?: Record<string, unknown>;
-    }) => (
-      <button data-testid="mui-icon-button" onClick={onClick}>
-        {children}
-      </button>
-    ),
-    InputAdornment: ({
-      children
-    }: {
-      children: ReactNode;
-      position: string;
-    }) => <span data-testid="mui-input-adornment">{children}</span>,
-    FormControlLabel: ({
-      control,
-      label
-    }: {
-      control: ReactNode;
-      label: ReactNode;
-    }) => (
-      <div data-testid="mui-form-control-label">
-        {control}
-        {label}
-      </div>
-    ),
-    Switch: ({
-      checked,
-      onChange,
-      disabled
-    }: {
-      checked?: boolean;
-      onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-      disabled?: boolean;
-    }) => (
-      <input
-        type="checkbox"
-        role="switch"
-        checked={checked}
-        onChange={onChange}
-        disabled={disabled}
-        data-testid="mui-switch"
-      />
-    )
-  };
-});
-
-// Mock per Material UI icons
-vi.mock('@mui/icons-material/RemoveCircleOutline', () => ({
-  default: () => <span data-testid="RemoveCircleOutlineIcon">RemoveIcon</span>
-}));
-
-// Type per aiutare con i test
-type TestFormValues = {
-  installments: Array<{
-    amount: string;
-    dueDate: Date | null;
-    isMultibeneficiary: boolean;
-  }>;
-};
-
-describe('InstallmentItem Component', () => {
-  // Valori comuni per tutti i test
-  const mockIndex = 0;
-  const mockField = { id: 'test-id' };
-  const mockFieldNamePrefix = 'installments';
-  const mockValidators: InstallmentValidators = {
-    amount: {
-      required: {
-        value: true,
-        message: 'debtPositionCreateWizard.step3.installments.amount.required'
-      },
-      validate: {
-        positive: (value: string) => {
-          if (!value) return true;
-          const numValue = parseFloat(value);
-          return (
-            numValue > 0 ||
-            'debtPositionCreateWizard.step3.installments.amount.positive'
-          );
-        },
-        validNumber: (value: string) => {
-          if (!value) return true;
-          return (
-            !isNaN(parseFloat(value)) ||
-            'debtPositionCreateWizard.step3.installments.amount.validNumber'
-          );
-        }
-      }
-    },
-    dueDate: {
-      required: 'debtPositionCreateWizard.step3.installments.dueDate.required'
-    }
-  };
-
-  // Mock delle funzioni e controlli
-  let mockControl: Control<TestFormValues>;
-  let mockErrors: FieldErrors<TestFormValues>;
-  let mockTrigger: UseFormTrigger<TestFormValues>;
-  let mockGetValues: UseFormGetValues<TestFormValues>;
-  let mockSetValue: UseFormSetValue<TestFormValues>;
-  let mockOnRemove: (index: number) => void;
-
-  beforeEach(() => {
-    // Reset dei mock
-    vi.clearAllMocks();
-
-    mockControl = {
-      _formValues: {
-        installments: [
-          { amount: '100', dueDate: null, isMultibeneficiary: false }
-        ]
-      }
-    } as unknown as Control<TestFormValues>;
-
-    mockErrors = {} as FieldErrors<TestFormValues>;
-
-    mockTrigger = vi.fn() as unknown as UseFormTrigger<TestFormValues>;
-
-    mockGetValues = vi.fn((path: Path<TestFormValues>): unknown => {
-      if (path === `${mockFieldNamePrefix}.${mockIndex}.amount`) {
-        return '100';
-      }
-      if (path === `${mockFieldNamePrefix}.${mockIndex}.dueDate`) {
-        return null;
-      }
-      if (path === `${mockFieldNamePrefix}.${mockIndex}.isMultibeneficiary`) {
-        return false;
-      }
-      return undefined;
-    }) as unknown as UseFormGetValues<TestFormValues>;
-
-    mockSetValue = vi.fn() as unknown as UseFormSetValue<TestFormValues>;
-
-    mockOnRemove = vi.fn();
-  });
-
-  it('renders installment component correctly', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-        onRemove={mockOnRemove}
-      />
-    );
-
-    // Verifica che i componenti principali siano renderizzati
-    expect(screen.getAllByTestId('mui-box').length).toBeGreaterThan(0);
-    expect(screen.getAllByTestId('mui-typography').length).toBeGreaterThan(0);
-    expect(screen.getAllByTestId('mui-textfield').length).toBeGreaterThan(0);
-    expect(screen.getByTestId('date-picker')).toBeInTheDocument();
-    expect(screen.getByTestId('RemoveCircleOutlineIcon')).toBeInTheDocument();
-  });
-
-  it('shows error message when amount is invalid and form is submitted', () => {
-    // Mock degli errori per il campo amount
-    mockErrors = {
-      installments: {
-        [mockIndex]: {
-          amount: {
-            type: 'required',
-            message:
-              'debtPositionCreateWizard.step3.installments.amount.required'
-          }
-        }
-      }
-    } as unknown as FieldErrors<TestFormValues>;
-
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={true}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-        onRemove={mockOnRemove}
-      />
-    );
-
-    // Cerchiamo tutti i TextField per vedere se mostrano errori
-    const textFields = screen.getAllByTestId('mui-textfield');
-    expect(textFields.length).toBeGreaterThan(0);
-
-    // Con l'implementazione attuale del mock è difficile verificare gli errori specifici,
-    // ma possiamo verificare che il componente sia stato renderizzato con gli errori
-    expect(mockErrors.installments?.[mockIndex]?.amount?.message).toBe(
-      'debtPositionCreateWizard.step3.installments.amount.required'
-    );
-  });
-
-  it('calls onRemove when remove button is clicked', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-        onRemove={mockOnRemove}
-      />
-    );
-
-    // Simula il click sul pulsante di rimozione
-    const removeButton = screen.getByTestId('mui-icon-button');
-    fireEvent.click(removeButton);
-
-    // Verifica che onRemove sia stato chiamato con l'indice corretto
-    expect(mockOnRemove).toHaveBeenCalledWith(mockIndex);
-  });
-
-  it('does not render remove button when onRemove is not provided', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-      />
-    );
-
-    // Dato che ora mockiamo in modo diverso, controlliamo il DOM manualmente
-    // per verificare che il pulsante non sia presente quando onRemove non è fornito
-    const removeButtons = screen.queryAllByTestId('mui-icon-button');
-    expect(removeButtons.length).toBe(0);
-  });
-
-  it('renders the switch for other beneficiaries', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-      />
-    );
-
-    // Verifica che lo switch sia renderizzato
-    expect(screen.getByTestId('mui-switch')).toBeInTheDocument();
-  });
-
-  it('disables all fields when disabled prop is true', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-        disabled={true}
-      />
-    );
-
-    // Verifichiamo che lo switch sia disabilitato
-    const switchElement = screen.getByTestId('mui-switch');
-    expect(switchElement).toHaveAttribute('disabled');
-  });
-
-  it('reports errors for dueDate field when submitted', () => {
-    // Mock degli errori per il campo dueDate
-    mockErrors = {
-      installments: {
-        [mockIndex]: {
-          dueDate: {
-            type: 'required',
-            message:
-              'debtPositionCreateWizard.step3.installments.dueDate.required'
-          }
-        }
-      }
-    } as unknown as FieldErrors<TestFormValues>;
-
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={true}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-      />
-    );
-
-    // Verifica che gli errori della data di scadenza siano passati correttamente
-    expect(mockErrors.installments?.[mockIndex]?.dueDate?.message).toBe(
-      'debtPositionCreateWizard.step3.installments.dueDate.required'
-    );
-  });
-
-  it('correctly handles due date changes', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-      />
-    );
-
-    // Trova il campo di input della data
-    const dateInput = screen.getByTestId('date-picker').querySelector('input');
-    expect(dateInput).toBeInTheDocument();
-
-    if (dateInput) {
-      // Simula il cambiamento della data
-      const newDate = '2025-01-01';
-      fireEvent.change(dateInput, { target: { value: newDate } });
-
-      // Non possiamo verificare direttamente l'effetto perché stiamo usando un mock
-      // ma possiamo verificare che l'input esista e sia interattivo
-    }
-  });
-
-  it('handles clearing the due date', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-      />
-    );
-
-    // Trova il pulsante di pulizia della data
-    const clearButton = screen.getByTestId('clear-date-button');
-    expect(clearButton).toBeInTheDocument();
-
-    // Simula il click per pulire la data
-    fireEvent.click(clearButton);
-
-    // Non possiamo verificare direttamente l'effetto perché stiamo usando un mock
-    // ma possiamo verificare che il pulsante esista e sia interattivo
-  });
-
-  it('can interact with input fields', () => {
-    render(
-      <InstallmentItem<TestFormValues>
-        index={mockIndex}
-        field={mockField}
-        control={mockControl}
-        errors={mockErrors}
-        isSubmitted={false}
-        validators={mockValidators}
-        fieldNamePrefix={mockFieldNamePrefix}
-        trigger={mockTrigger}
-        getValues={mockGetValues}
-        setValue={mockSetValue}
-      />
-    );
-
-    // Trova tutti gli input text per verificare che siano interattivi
-    const inputs = screen.getAllByTestId('textfield-input');
-    expect(inputs.length).toBeGreaterThan(0);
-
-    // Verifica che sia possibile interagire con l'input
-    const input = inputs[0];
-    fireEvent.change(input, { target: { value: '200' } });
-    fireEvent.blur(input);
-
-    // Verifica che lo switch sia interattivo
-    const switchElement = screen.getByTestId('mui-switch');
-    expect(switchElement).toBeInTheDocument();
-    fireEvent.click(switchElement);
-  });
-});
